requirements.txt
PyGithub
pydantic